requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
openai-agents>=0.2.10
